Modified data layer and enabled saving and loading WIP
