+++ src/LayerEditor/data/__init__.py
@@ -1,0 +1,3 @@
+from .config import _Config
+
+cfg = _Config.open()
