+++ src/LayerEditor/helpers/keyboard_shortcuts_definition.py
