+++ src/LayerEditor/layer_editor.py
@@ -6,7 +6,7 @@ import os
 import signal
 import argparse
 
-#from LayerEditor.data.geo_data import GeometryData
+from LayerEditor.data.geo_data import GeometryData
 
 sys.path.append(os.path.dirname(os.path.dirname(os.path.realpath(__file__))))
 
@@ -15,7 +15,7 @@ import PyQt5.QtWidgets as QtWidgets
 import PyQt5.QtCore as QtCore
 import gm_base.icon as icon
 from gm_base.geomop_util_qt import Autosave
-#from LayerEditor.data.config import _Config
+from LayerEditor.data.config import _Config
 
 
 class LayerEditor:
@@ -28,15 +28,15 @@ class LayerEditor:
         #    self.parse_args(args)
 
 
-        #self.geo_data = GeometryData()
-        #self.cfg = _Config.open()
+        self.geo_data = GeometryData()
+        self.cfg = _Config.open()
 
         self.mainwindow = MainWindow(self)
         self.exit = False
 
-        # save_fn = lambda c=self.cfg: self.geo_data.le_serializer.save(c)
-        # self.autosave = Autosave(self.cfg.current_workdir, lambda: self.cfg.curr_file, save_fn)
-        # self._restore_backup()
+        save_fn = lambda c=self.cfg: self.geo_data.le_serializer.save(c)
+        self.autosave = Autosave(self.cfg.current_workdir, lambda: self.cfg.curr_file, save_fn)
+        self._restore_backup()
 
         # show
         self.mainwindow.show()
@@ -59,7 +59,7 @@ class LayerEditor:
     #     Parse cmd line args.
     #     :return:
     #     """
-    #     parser = argparse.ArgumentParser(description='LayerEditor')
+    #     parser = argparse.ArgumentParser(description='OldLayerEditor')
     #     parser.add_argument('--debug', action='store_true', help='Enable debug mode')
     #     parser.add_argument('file', help='Layers geometry JSON file.', default=None, nargs='?')
     #     args = parser.parse_args(args)
@@ -101,52 +101,66 @@ class LayerEditor:
     #     #
     #     log_unhandled_exceptions(cfg.config.__class__.CONTEXT_NAME, on_unhandled_exception)
     #
-    # def _restore_backup(self):
-    #     """recover file from backup file if it exists and if user wishes so"""
-    #     restored = self.autosave.restore_backup()
-    #     if restored:
-    #         cfg.main_window.release_data(cfg.diagram_id())
-    #         cfg.history.remove_all()
-    #         cfg.le_serializer.load(cfg, self.autosave.backup_filename())
-    #         cfg.main_window.refresh_all()
-    #         cfg.history.last_save_labels = -1
-    #     self.autosave.autosave_timer.stop()
-    #     return restored
-    #
-    # def new_file(self):
-    #     """
-    #     New file action handler.
-    #     return close if no file is created nor openned.
-    #     """
-    #     if not self.save_old_file():
-    #         return
-    #
-    #     cfg.new_file()
-    #     cfg.diagram.area.set_area([(0, 0), (100, 0), (100, 100), (0, 100)])
-    #     self.mainwindow.refresh_all()
-    #     self.mainwindow.paint_new_data()
-    #     self._update_document_name()
-    #     return True
-    #
-    # def open_file(self, in_file=None):
-    #     """
-    #     open file menu action
-    #     handler for triggered signal.
-    #     """
-    #
-    #     if not self.save_old_file():
-    #         return False
-    #     if in_file is None:
-    #         in_file, _ = QtWidgets.QFileDialog.getOpenFileName(
-    #             self.mainwindow, "Choose Json Geometry File",
-    #             cfg.config.data_dir, "Json Files (*.json)")
-    #     if in_file:
-    #         cfg.open_file(in_file)
-    #         self._update_document_name()
-    #         self._restore_backup()
-    #         self.mainwindow.show_status_message("File {} is opened".format(in_file))
-    #         return True
-    #     return False
+    def _restore_backup(self):
+        """recover file from backup file if it exists and if user wishes so"""
+        restored = self.autosave.restore_backup()
+        if restored:
+            #self.cfg.main_window.release_data(self.cfg.diagram_id())
+            #self.cfg.history.remove_all()
+            self.geo_data.le_serializer.load(self.cfg, self.autosave.backup_filename())
+            #self.cfg.main_window.refresh_all()
+            #self.cfg.history.last_save_labels = -1
+        self.autosave.autosave_timer.stop()
+        return restored
+
+    def new_file(self):
+        """
+        New file action handler.
+        return close if no file is created nor openned.
+        """
+        if not self.save_old_file():
+            return
+
+        #self.mainwindow.release_data(cls.diagram_id())
+        self.geo_data = GeometryData()
+
+        self.geo_data.curr_diagram.area.set_area([(0, 0), (100, 0), (100, 100), (0, 100)])
+        #self.mainwindow.refresh_all()
+        #self.mainwindow.paint_new_data()
+        self._update_document_name()
+        return True
+
+    def open_file(self, in_file=None):
+        """
+        open file menu action
+        handler for triggered signal.
+        """
+
+        if not self.save_old_file():
+            return False
+        if in_file is None:
+            in_file, _ = QtWidgets.QFileDialog.getOpenFileName(
+                self.mainwindow, "Choose Json Geometry File",
+                self.cfg.data_dir, "Json Files (*.json)")
+        if in_file:
+            #self.main_window.release_data(cls.diagram_id())
+            self.geo_data = GeometryData()
+            self.geo_data.curr_file = in_file
+            self.cfg.update_current_workdir(in_file)
+            if self.geo_data.curr_file is not None:
+                try:
+                    self.geo_data.curr_file_timestamp = os.path.getmtime(in_file)
+                except OSError:
+                    pass
+
+            self.geo_data.load(in_file)
+            #cls.main_window.refresh_all()
+            self.cfg.add_recent_file(in_file)
+            self._update_document_name()
+            self._restore_backup()
+            self.mainwindow.show_status_message("File {} is opened".format(in_file))
+            return True
+        return False
     #
     # def add_shape_file(self):
     #     """open set file"""
@@ -178,79 +192,81 @@ class LayerEditor:
     #     self._restore_backup()
     #     self.mainwindow.show_status_message("File '" + action.data() + "' is opened")
     #
-    # def save_file(self):
-    #     """save file menu action"""
-    #     if cfg.curr_file is None:
-    #         return self.save_as()
-    #     if cfg.confront_file_timestamp():
-    #         return
-    #     cfg.save_file()
-    #     self.autosave.delete_backup()
-    #     self.mainwindow.show_status_message("File is saved")
-    #
-    # def save_as(self):
-    #     """save file menu action"""
-    #     if cfg.confront_file_timestamp():
-    #         return
-    #     if cfg.curr_file is None:
-    #         new_file = cfg.config.data_dir + os.path.sep + "NewFile.json"
-    #     else:
-    #         new_file = cfg.curr_file
-    #     dialog = QtWidgets.QFileDialog(self.mainwindow, 'Save as JSON Geometry File', new_file,
-    #                                    "JSON Files (*.json)")
-    #     dialog.setDefaultSuffix('.json')
-    #     dialog.setFileMode(QtWidgets.QFileDialog.AnyFile)
-    #     dialog.setAcceptMode(QtWidgets.QFileDialog.AcceptSave)
-    #     dialog.setOption(QtWidgets.QFileDialog.DontConfirmOverwrite, False)
-    #     dialog.setViewMode(QtWidgets.QFileDialog.Detail)
-    #     if dialog.exec_():
-    #         self.autosave.delete_backup()
-    #         file_name = dialog.selectedFiles()[0]
-    #         cfg.save_file(file_name)
-    #         self.mainwindow.update_recent_files()
-    #         self._update_document_name()
-    #         self.mainwindow.show_status_message("File is saved")
-    #         return True
-    #     return False
-    #
-    # def save_old_file(self):
-    #     """
-    #     If file not saved, display confirmation dialog and if is needed, do it
-    #
-    #     return: False if action have to be aborted
-    #     """
-    #     if cfg.changed():
-    #         msg_box = QtWidgets.QMessageBox()
-    #         msg_box.setWindowTitle("Confirmation")
-    #         msg_box.setIcon(QtWidgets.QMessageBox.Question)
-    #         msg_box.setStandardButtons(QtWidgets.QMessageBox.Yes | QtWidgets.QMessageBox.No | QtWidgets.QMessageBox.Abort)
-    #         msg_box.setDefaultButton(QtWidgets.QMessageBox.Abort)
-    #         msg_box.setText("The document has unsaved changes, do you want to save it?")
-    #         reply = msg_box.exec_()
-    #
-    #         if reply == QtWidgets.QMessageBox.Abort:
-    #             return False
-    #         if reply == QtWidgets.QMessageBox.Yes:
-    #             if cfg.curr_file is None:
-    #                 return self.save_as()
-    #             else:
-    #                 self.save_file()
-    #         if reply == QtWidgets.QMessageBox.No:
-    #             self.autosave.delete_backup()
-    #     return True
-    #
+    def save_file(self):
+        """save file menu action"""
+        if self.cfg.curr_file is None:
+            return self.save_as()
+        if self.cfg.confront_file_timestamp():
+            return
+        self.geo_data.save_file()
+        self.cfg.add_recent_file(self.geo_data.curr_file)
+        self.autosave.delete_backup()
+        self.mainwindow.show_status_message("File is saved")
+
+    def save_as(self):
+        """save file menu action"""
+        if self.geo_data.confront_file_timestamp():
+            return
+        if self.geo_data.curr_file is None:
+            new_file = self.cfg.data_dir + os.path.sep + "NewFile.json"
+        else:
+            new_file = self.geo_data.curr_file
+        dialog = QtWidgets.QFileDialog(self.mainwindow, 'Save as JSON Geometry File', new_file,
+                                       "JSON Files (*.json)")
+        dialog.setDefaultSuffix('.json')
+        dialog.setFileMode(QtWidgets.QFileDialog.AnyFile)
+        dialog.setAcceptMode(QtWidgets.QFileDialog.AcceptSave)
+        dialog.setOption(QtWidgets.QFileDialog.DontConfirmOverwrite, False)
+        dialog.setViewMode(QtWidgets.QFileDialog.Detail)
+        if dialog.exec_():
+            self.autosave.delete_backup()
+            file_name = dialog.selectedFiles()[0]
+            self.geo_data.save_file(file_name)
+            self.cfg.add_recent_file(file_name)
+            self.mainwindow.update_recent_files()
+            self._update_document_name()
+            self.mainwindow.show_status_message("File is saved")
+            return True
+        return False
+
+    def save_old_file(self):
+        """
+        If file not saved, display confirmation dialog and if is needed, do it
+
+        return: False if action have to be aborted
+        """
+        if self.cfg.changed():
+            msg_box = QtWidgets.QMessageBox()
+            msg_box.setWindowTitle("Confirmation")
+            msg_box.setIcon(QtWidgets.QMessageBox.Question)
+            msg_box.setStandardButtons(QtWidgets.QMessageBox.Yes | QtWidgets.QMessageBox.No | QtWidgets.QMessageBox.Abort)
+            msg_box.setDefaultButton(QtWidgets.QMessageBox.Abort)
+            msg_box.setText("The document has unsaved changes, do you want to save it?")
+            reply = msg_box.exec_()
+
+            if reply == QtWidgets.QMessageBox.Abort:
+                return False
+            if reply == QtWidgets.QMessageBox.Yes:
+                if self.cfg.curr_file is None:
+                    return self.save_as()
+                else:
+                    self.save_file()
+            if reply == QtWidgets.QMessageBox.No:
+                self.autosave.delete_backup()
+        return True
+
     # def run(self):
     #     """go"""
     #     self._app.exec()
     #
-    # def _update_document_name(self):
-    #     """Update document title (add file name)"""
-    #     title = "GeoMop Layer Editor"
-    #     if cfg.curr_file is None:
-    #         title += " - New File"
-    #     else:
-    #         title += " - " + cfg.curr_file
-    #     self.mainwindow.setWindowTitle(title)
+    def _update_document_name(self):
+        """Update document title (add file name)"""
+        title = "GeoMop Layer Editor"
+        if self.cfg.curr_file is None:
+            title += " - New File"
+        else:
+            title += " - " + self.cfg.curr_file
+        self.mainwindow.setWindowTitle(title)
 
 
 if __name__ == "__main__":
