--- src/LayerEditor/leconfig/__init__.py
@@ -1,5 +1,0 @@
-"""
-.. codeauthor:: Pavel Richter <pavel.richter@tul.cz>
-"""
-
-from .leconfig import LEConfig as cfg
